+++ modules/meteroid/crates/meteroid-store/src/services/quotes.rs
@@ -117,7 +117,7 @@ fn build_subscription_from_quote(
             fee: a.fee.clone(),
             product_id: a.product_id,
             price_id: a.price_id,
-            quantity: a.quantity as u32,
+            quantity: a.quantity,
         })
         .collect();
